Add `force-vandusen-notification?' keyword parameter for `pastiche'
Default value: #f.

When set to a truthy value, the "Please notify the #chicken channel on
freenode." checkbox won't be displayed and vandusen will
unconditionally be notified.

This feature is targeted at paste.call-cc.org, which has been used by
spammers from time to time.  Spammers usually uncheck the "notify
vandusen" checkbox, which makes it difficult to track spam pastes.
With this feature, we will at least be notified on #chicken.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>